RSpecs working in the blog. 6/6 specs succeeded. (Including model and controller specs)
This is a combination of 5 commits.

The first commit's message is:
branch (Commit used to solve troubles in FactoryGirl.reload)

This is the 2nd commit message:
spec for ArticlesController created. Context1: User logged in. Context 2: User logged out, redirect. TO SOLVE: FactoryGirl in userLoggedIn and Locale+redirect in usedLoggedOut

This is the 3rd commit message:
fixes in specs

This is the 4th commit message:
change the location of factories. Convention over  configuration

 This is the 5th commit message:
specs created. 6/6 specs succeeded